feat: isolate the retrieval of the active user in a utility and inject the active user into the change password form
