+++ src/app/_features/auth/utils.ts
@@ -1,5 +1,9 @@
-import { KEY } from '@/app/_features/auth/constants';
+import UserRepository from '@/app/_data/user';
+import { signOut } from '@/app/_features/auth/actions';
+import { KEY, SESSION_ID_COOKIE } from '@/app/_features/auth/constants';
+import { type User } from '@prisma/client';
 import { SignJWT } from 'jose';
+import { cookies } from 'next/headers';
 
 /**
  * Generates a token.
@@ -11,4 +15,23 @@ export async function generateToken(): Promise<string> {
     .setIssuedAt()
     .setExpirationTime('1day')
     .sign(KEY);
+}
+
+/**
+ * Gets the user associated with the active session.
+ * @returns {Promise<User>} The user.
+ */
+export async function getSessionUser(): Promise<User> {
+  const cookieStore = await cookies();
+
+  const sessionId = cookieStore.get(SESSION_ID_COOKIE.name);
+
+  const user = await UserRepository.findBySession(sessionId?.value as string);
+
+  if (!user) {
+    await signOut();
+    throw new Error('User not found for active session');
+  }
+
+  return user;
 }
