+++ src/app/change-password/page.tsx
@@ -1,12 +1,15 @@
+import { getSessionUser } from '@/app/_features/auth/utils';
 import ChangePasswordForm from '@/app/_features/change-password/components/change-password-form';
 
 /**
  * Change password page component.
  */
-export default function ChangePassword() {
+export default async function ChangePassword() {
+  const user = await getSessionUser();
+
   return (
     <main>
-      <ChangePasswordForm />
+      <ChangePasswordForm user={user} />
     </main>
   );
 }
